Redo SiegeCommand
